requirements.txt
click==8.1.1
et-xmlfile==1.1.0
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
openpyxl==3.0.9
python-dotenv==0.20.0
SQLAlchemy==1.4.33
Werkzeug==2.1.0
zipp==3.7.0
